+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - [breaking]: `.number()` method was renamed to `.toString()`;
 - [feature]: ranks custom decoding options;
 - [breaking]: removed aliases for `.setRadix()` method properties;
+- [breaking]: renamed `.asDecimal` property to `.decimal`;
 
 # 0.3.0 (2021-08-28)